@@ -1,5 +1,13 @@
 module github.com/Mugen-Builders/to-do-memory
 
-go 1.20
+go 1.23.0
 
-require github.com/google/wire v0.6.0
+toolchain go1.23.7
+
+require github.com/ethereum/go-ethereum v1.15.5
+
+require (
+	github.com/holiman/uint256 v1.3.2 // indirect
+	golang.org/x/crypto v0.32.0 // indirect
+	golang.org/x/sys v0.29.0 // indirect
+)
