@@ -1,16 +1,21 @@
 module github.com/Mugen-Builders/to-do-sqlite
 
-go 1.20
+go 1.23.0
+
+toolchain go1.23.7
 
 require (
-	github.com/google/wire v0.6.0
+	github.com/ethereum/go-ethereum v1.15.5
 	gorm.io/driver/sqlite v1.5.7
 	gorm.io/gorm v1.25.12
 )
 
 require (
+	github.com/holiman/uint256 v1.3.2 // indirect
 	github.com/jinzhu/inflection v1.0.0 // indirect
 	github.com/jinzhu/now v1.1.5 // indirect
 	github.com/mattn/go-sqlite3 v1.14.22 // indirect
-	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/crypto v0.32.0 // indirect
+	golang.org/x/sys v0.29.0 // indirect
+	golang.org/x/text v0.21.0 // indirect
 )
